convert_meshes_requirements.txt
numpy==2.1.3
pillow==11.0.0
pycollada==0.8
python-dateutil==2.9.0.post0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.16.0
trimesh==4.5.3
